backend/requirements.txt
fastapi
uvicorn
supabase
anthropic
python-multipart
python-dotenv
reportlab
qrcode
pillow